requirements.txt
# Core dependencies
PyQt6>=6.5.0
anthropic>=0.18.0
langchain>=0.1.0
langchain_text_splitters>=0.0.1
PyPDF2>=3.0.0
azure-ai-documentintelligence>=1.0.0
azure-core>=1.28.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
